Repackage python modules as separate packages.
This repackaging allows for Python 2 and Python 3 versions of the
youtube_dl modules to co-exist. The youtube-dl program will from now
on be executed under Python 3 and will require the python3-youtube-dl
package. The python2-youtube-dl package may be installed, in addition
to these or even by itself, in case Python 2 programs need to make use
of the youtube_dl modules.

Closes: #723990